Fix type mismatch, nsdpy.h is a standard interface, TODO investigate if other renderers use size_t or int.

+++ displays/d_sdcWin32/interface.c
@@ -26,7 +26,7 @@
 
 PtDspyError DspyImageQuery(PtDspyImageHandle image,
                            PtDspyQueryType   type,
-                           size_t            size,
+                           int	             size,
                            void              *data)
 {
 #if SHOW_CALLSTACK
